# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plateforme E-learning",
        "description": "Documentation API pour la plateforme e-learning",
        "contact": {
            "email": "diopawa425@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API Local"
        }
    ],
    "paths": {
...
        "/api/courses/{course}/lessons": {
            "get": {
                "tags": [
                    "Leçon"
                ],
                "summary": "Lister tous les lessons",
                "operationId": "2c4e6971ea10678972bd235206d31c64",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des Lessosn",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Lesson"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Leçon"
                ],
                "summary": "Création de nouvelle leçon",
                "operationId": "0a1fc3a4f8333d9bc03a9ba84b620eb4",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Lesson"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Leçon créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/courses/{course}/lessons/{lesson}": {
            "get": {
                "tags": [
                    "Leçon"
                ],
                "summary": "Detail lesson",
                "operationId": "8c11087adc5be274079d268998a95771",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lesson",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/lessons/{lesson}": {
            "put": {
                "tags": [
                    "Leçon"
                ],
                "summary": "Modifier ou mettre a jour les inlfos d'une Leçon",
                "operationId": "73b130595c873899cec53486a9964352",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Lesson"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Modifier avec succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lesson non trougvé"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Leçon"
                ],
                "summary": "Supprimer une Leçon",
                "operationId": "9cebf75653bf2011b4b9d952a587217d",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID de la Leçon à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Leçon supprimé avec succès"
                    },
                    "404": {
                        "description": "Leçon non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Content": {
                "required": [
                    "id",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "lesson_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "type": {
                        "type": "string",
                        "example": "pdf"
                    },
                    "data": {
                        "type": "string",
                        "nullable": true
                    },
                    "file_path": {
                        "type": "string",
                        "example": "lessons/files/example.pdf"
                    },
                    "external_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Lesson": {
                "required": [
                    "id",
                    "title",
                    "order",
                    "duration_minutes"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction à Laravel"
                    },
                    "course_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    },
                    "duration_minutes": {
                        "type": "integer",
                        "example": 60
                    },
                    "is_published": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_locked": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "contents": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Content"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Entrez le token Sanctum reçu après connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}